pyproject.toml
# ── Project ───────────────────────────────────────────────
[project]
name = "data-engineering-notes"
version = "0.1.0"
description = ""
authors = [
    {name = "Camille",email = "camille.larpin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "anthropic (>=0.84.0,<0.85.0)",
    "click (>=8.3.1,<9.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-telegram-bot (>=22.6,<23.0)"
]

# ── Build ──────────────────────────────────────────────────
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

# ── Dependencies ───────────────────────────────────────────
[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]

# ── Testing ────────────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["."]

# ── Linting & Formatting ───────────────────────────────────
[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N"]  # style, pyflakes, isort, naming
